# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobCreateRequest": {
        "description": "Create a background simulation job.",
        "properties": {
          "height": {
            "default": 64,
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Height",
            "type": "integer"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 4294967295.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "steps": {
            "default": 10,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Steps",
            "type": "integer"
          },
          "width": {
            "default": 64,
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "title": "JobCreateRequest",
        "type": "object"
      },
      "JobSummary": {
        "description": "Job status payload.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at",
          "request"
        ],
        "title": "JobSummary",
        "type": "object"
      },
...
      "PresetDetail": {
        "description": "Preset detail including portable psyfi parameter defaults.",
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "coupling_strength": {
            "title": "Coupling Strength",
            "type": "number"
          },
          "drift_amplitude": {
            "title": "Drift Amplitude",
            "type": "number"
          },
          "drift_velocity": {
            "title": "Drift Velocity",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "normalization_P": {
            "title": "Normalization P",
            "type": "number"
          },
          "normalization_V": {
            "title": "Normalization V",
            "type": "number"
          },
          "phase_noise": {
            "title": "Phase Noise",
            "type": "number"
          },
          "phase_reset_strength": {
            "title": "Phase Reset Strength",
            "type": "number"
          },
          "substance_class": {
            "title": "Substance Class",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "substance_class",
          "coupling_strength",
          "phase_noise",
          "phase_reset_strength",
          "drift_amplitude",
          "drift_velocity",
          "normalization_P",
          "normalization_V"
        ],
        "title": "PresetDetail",
        "type": "object"
      },
...
      "PsyFiSession": {
        "description": "Deterministic session document shared by web and future iOS clients.",
        "properties": {
          "abx_core": {
            "default": "1.3",
            "title": "Abx Core",
            "type": "string"
          },
          "api_version": {
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "engine_version": {
            "default": "0.1.0",
            "title": "Engine Version",
            "type": "string"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Labels",
            "type": "object"
          },
          "parameters": {
            "$ref": "#/components/schemas/SessionParameters"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "provenance": {
            "$ref": "#/components/schemas/SessionProvenance"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "default": "psyfi.session.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "seed": {
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "seed",
          "parameters",
          "provenance"
        ],
        "title": "PsyFiSession",
        "type": "object"
      },
      "PsyFiVisualization": {
        "description": "Renderer-independent visualization schema (Canvas/WebGL/WebGPU/Metal).",
        "properties": {
          "accessibility": {
            "$ref": "#/components/schemas/VisualizationAccessibility"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/VisualizationChannel"
            },
            "title": "Channels",
            "type": "array"
          },
          "field": {
            "additionalProperties": true,
            "title": "Field",
            "type": "object"
          },
          "frame_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Index"
          },
          "layers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Layers",
            "type": "array"
          },
          "provenance_id": {
            "title": "Provenance Id",
            "type": "string"
          },
          "renderer_requirements": {
            "additionalProperties": true,
            "title": "Renderer Requirements",
            "type": "object"
          },
          "schema_version": {
            "default": "psyfi.visualization.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "sequence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence Id"
          },
          "view": {
            "additionalProperties": true,
            "title": "View",
            "type": "object"
          }
        },
        "required": [
          "provenance_id",
          "field",
          "channels",
          "accessibility"
        ],
        "title": "PsyFiVisualization",
        "type": "object"
      },
...
      "SessionParameters": {
        "description": "Canonical parameters required to reproduce a simulation.",
        "properties": {
          "coupling_strength": {
            "default": 0.5,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Coupling Strength",
            "type": "number"
          },
          "drift_amplitude": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drift Amplitude"
          },
          "drift_velocity": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drift Velocity"
          },
          "height": {
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Height",
            "type": "integer"
          },
          "normalization_P": {
            "default": 1.0,
            "title": "Normalization P",
            "type": "number"
          },
          "normalization_V": {
            "default": 1.0,
            "title": "Normalization V",
            "type": "number"
          },
          "phase_noise": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase Noise"
          },
          "phase_reset_strength": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase Reset Strength"
          },
          "steps": {
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Steps",
            "type": "integer"
          },
          "width": {
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "width",
          "height",
          "steps"
        ],
        "title": "SessionParameters",
        "type": "object"
      },
      "SessionProvenance": {
        "description": "Portable provenance subset derived from ABX ProvenanceRecord.",
        "properties": {
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "module_chain": {
            "items": {
              "type": "string"
            },
            "title": "Module Chain",
            "type": "array"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "id"
        ],
        "title": "SessionProvenance",
        "type": "object"
      },
...
      "SimulateRequest": {
        "description": "Request for consciousness field simulation.",
        "properties": {
          "height": {
            "default": 64,
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Height",
            "type": "integer"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional substance preset id/alias influencing coupling and normalization.",
            "title": "Preset"
          },
          "seed": {
            "anyOf": [
              {
                "maximum": 4294967295.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "steps": {
            "default": 10,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Steps",
            "type": "integer"
          },
          "width": {
            "default": 64,
            "maximum": 512.0,
            "minimum": 8.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "title": "SimulateRequest",
        "type": "object"
      },
      "SimulateResponse": {
        "description": "Response from consciousness field simulation.",
        "properties": {
          "api_version": {
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "coherence": {
            "title": "Coherence",
            "type": "number"
          },
          "engine_version": {
            "default": "0.1.0",
            "title": "Engine Version",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "module_chain": {
            "items": {
              "type": "string"
            },
            "title": "Module Chain",
            "type": "array"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "provenance_id": {
            "title": "Provenance Id",
            "type": "string"
          },
          "richness": {
            "title": "Richness",
            "type": "number"
          },
          "roughness": {
            "title": "Roughness",
            "type": "number"
          },
          "schema_version": {
            "default": "psyfi.session.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "session": {
            "$ref": "#/components/schemas/PsyFiSession"
          },
          "symmetry": {
            "title": "Symmetry",
            "type": "number"
          },
          "valence": {
            "title": "Valence",
            "type": "number"
          },
          "visualization": {
            "$ref": "#/components/schemas/PsyFiVisualization"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "width",
          "height",
          "valence",
          "coherence",
          "symmetry",
          "roughness",
          "richness",
          "seed",
          "provenance_id",
          "module_chain",
          "session",
          "visualization"
        ],
        "title": "SimulateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Consciousness field simulation engine with ABX-Core v1.3",
    "title": "PsyFi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/simulate": {
      "post": {
        "description": "Start a cancellable simulation job in a worker thread.",
        "operationId": "create_simulate_job_api_jobs_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Simulate Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "description": "Request cancellation of a queued/running job.",
        "operationId": "cancel_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "Fetch job status/result.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/presets/{preset_id}": {
      "get": {
        "description": "Fetch one preset by id or alias.",
        "operationId": "get_substance_preset_api_presets__preset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Substance Preset",
        "tags": [
          "presets"
        ]
      }
    },
...
    "/api/v1/jobs/simulate": {
      "post": {
        "description": "Start a cancellable simulation job in a worker thread.",
        "operationId": "create_simulate_job_api_v1_jobs_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Simulate Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "delete": {
        "description": "Request cancellation of a queued/running job.",
        "operationId": "cancel_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "Fetch job status/result.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/presets/{preset_id}": {
      "get": {
        "description": "Fetch one preset by id or alias.",
        "operationId": "get_substance_preset_api_v1_presets__preset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Substance Preset",
        "tags": [
          "presets"
        ]
      }
    },
...
    "/api/v1/simulate/": {
      "post": {
        "description": "Simulate consciousness field evolution (cancellable on disconnect).",
        "operationId": "simulate_consciousness_field_api_v1_simulate__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Consciousness Field",
        "tags": [
          "simulation"
        ]
      }
    },
...
    "/simulate/": {
      "post": {
        "description": "Simulate consciousness field evolution (cancellable on disconnect).",
        "operationId": "simulate_consciousness_field_simulate__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Consciousness Field",
        "tags": [
          "simulation"
        ]
      }
    }
  }
}